+++ addon/utils/general/keyboard.js
@@ -1,0 +1,14 @@
+export const keys = {
+  8: 'backspace',
+  9: 'tab',
+  13: 'enter',
+  27: 'escape',
+  37: 'left',
+  38: 'up',
+  39: 'right',
+  40: 'down'
+};
+
+export function getKeyName(e) {
+  return keys[e.keyCode];
+}