+++ tests/integration/components/select-box/activating-options-test.js
@@ -149,6 +149,11 @@ module('select-box (activating options)', function (hooks) {
   test('activation boundaries', async function (assert) {
     assert.expect(8);
 
+    // We specifically must use keydown not keyup, because
+    // this allows us (or the user of the addon) to preventDefault on
+    // the event. Thereby stopping the page from scrolling when trying
+    // to navigate.
+
     this.handlePressDown = (e, sb) => sb.activateNextOption();
     this.handlePressUp = (e, sb) => sb.activatePreviousOption();
 
@@ -167,19 +172,19 @@ module('select-box (activating options)', function (hooks) {
       .dom('.select-box__option[aria-current="true"]')
       .doesNotExist('precondition: nothing active');
 
-    await triggerKeyEvent('.select-box', 'keyup', 40); // Down
+    await triggerKeyEvent('.select-box', 'keydown', 40); // Down
 
     assert.dom('.select-box__option[aria-current="true"]').hasText('One');
 
-    await triggerKeyEvent('.select-box', 'keyup', 40); // Down
+    await triggerKeyEvent('.select-box', 'keydown', 40); // Down
 
     assert.dom('.select-box__option[aria-current="true"]').hasText('Two');
 
-    await triggerKeyEvent('.select-box', 'keyup', 40); // Down
+    await triggerKeyEvent('.select-box', 'keydown', 40); // Down
 
     assert.dom('.select-box__option[aria-current="true"]').hasText('Three');
 
-    await triggerKeyEvent('.select-box', 'keyup', 40); // Down
+    await triggerKeyEvent('.select-box', 'keydown', 40); // Down
 
     assert
       .dom('.select-box__option[aria-current="true"]')
@@ -188,15 +193,15 @@ module('select-box (activating options)', function (hooks) {
         'does not cycle back to the beginning when reaching the end'
       );
 
-    await triggerKeyEvent('.select-box', 'keyup', 38); // Up
+    await triggerKeyEvent('.select-box', 'keydown', 38); // Up
 
     assert.dom('.select-box__option[aria-current="true"]').hasText('Two');
 
-    await triggerKeyEvent('.select-box', 'keyup', 38); // Up
+    await triggerKeyEvent('.select-box', 'keydown', 38); // Up
 
     assert.dom('.select-box__option[aria-current="true"]').hasText('One');
 
-    await triggerKeyEvent('.select-box', 'keyup', 38); // Up
+    await triggerKeyEvent('.select-box', 'keydown', 38); // Up
 
     assert
       .dom('.select-box__option[aria-current="true"]')